+++ R/checkArgumentsFragPipe.R
@@ -9,7 +9,7 @@
     forceOverwrite, experimentInfo, species, fragpipeDir,
     idCol, labelCol, geneIdCol, proteinIdCol, stringIdCol, extraFeatureCols,
     iColPattern, sampleAnnot, includeOnlySamples, excludeSamples,
-    imputeMethod, assaysForExport, addAbundanceValues,
+    imputeMethod, imputeArgs, assaysForExport, addAbundanceValues,
     addHeatmaps, mergeGroups, filtersSE,
     comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
@@ -152,7 +152,10 @@
 
     ## Method choices
     .assertScalar(x = imputeMethod, type = "character",
-                  validValues = c("impSeqRob", "MinProb", "MinProbGlobal"))
+                  validValues = c("impSeqRob", "MinProb", "MinProbGlobal",
+                                  "MinProbOffset", "MinProbGlobalOffset",
+                                  "custom"))
+    .assertVector(x = imputeArgs, type = "list")
     .assertVector(x = assaysForExport, type = "character", allowNULL = TRUE)
     .assertScalar(x = addAbundanceValues, type = "logical")
     .assertScalar(x = addHeatmaps, type = "logical")