+++ R/checkArgumentsSE.R
@@ -10,7 +10,7 @@
         templateRmd, outputDir, outputBaseName, reportTitle, reportAuthor,
         forceOverwrite, experimentInfo, species, inFile, summaryInfo, aName,
         idCol, labelCol, geneIdCol, proteinIdCol, stringIdCol, extraFeatureCols,
-        imputeMethod, assaysForExport, addAbundanceValues,
+        imputeMethod, imputeArgs, assaysForExport, addAbundanceValues,
         addHeatmaps, mergeGroups, filtersSE,
         comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
         subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
@@ -119,7 +119,10 @@
 
     ## Method choices
     .assertScalar(x = imputeMethod, type = "character",
-                  validValues = c("impSeqRob", "MinProb", "MinProbGlobal"))
+                  validValues = c("impSeqRob", "MinProb", "MinProbGlobal",
+                                  "MinProbOffset", "MinProbGlobalOffset",
+                                  "custom"))
+    .assertVector(x = imputeArgs, type = "list")
     .assertVector(x = assaysForExport, type = "character", allowNULL = TRUE)
     .assertScalar(x = addAbundanceValues, type = "logical")
     .assertScalar(x = addHeatmaps, type = "logical")