+++ tests/testthat/test-checkArgumentsDIANN.R
@@ -37,6 +37,7 @@ test_that("argument checking for DIANN works", {
         filtersDF = list(),
         filtersSE = list(),
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = NULL,
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -346,6 +347,12 @@ test_that("argument checking for DIANN works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(.checkArgumentsDIANN, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(.checkArgumentsDIANN, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0