+++ tests/testthat/test-checkArgumentsMaxQuant.R
@@ -41,6 +41,7 @@ test_that("argument checking for MQ works", {
         excludeSamples = "",
         filtersSE = einprotMQFilters,
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = c("iBAQ", "Top3"),
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -324,6 +325,12 @@ test_that("argument checking for MQ works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(.checkArgumentsMaxQuant, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(.checkArgumentsMaxQuant, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0