+++ tests/testthat/test-checkArgumentsPDTMT.R
@@ -47,6 +47,7 @@ test_that("argument checking for PD-TMT works", {
         excludeSamples = "",
         filtersSE = einprotPDTMTProteinFilters,
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = NULL,
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -397,6 +398,12 @@ test_that("argument checking for PD-TMT works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(.checkArgumentsPDTMT, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(.checkArgumentsPDTMT, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0