+++ tests/testthat/test-checkArgumentsSpectronaut.R
@@ -45,6 +45,7 @@ test_that("argument checking for Spectronaut works", {
         filtersDF = list(),
         filtersSE = list(),
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = NULL,
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -365,6 +366,12 @@ test_that("argument checking for Spectronaut works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(.checkArgumentsSpectronaut, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(.checkArgumentsSpectronaut, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0