+++ tests/testthat/test-runPDTMTAnalysis.R
@@ -46,6 +46,7 @@ test_that("runPDTMTAnalysis works", {
         excludeSamples = "",
         filtersSE = einprotPDTMTProteinFilters,
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = NULL,
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -373,6 +374,12 @@ test_that("runPDTMTAnalysis works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(runPDTMTAnalysis, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0