+++ tests/testthat/test-runSpectronautAnalysis.R
@@ -46,6 +46,7 @@ test_that("runSpectronautAnalysis works", {
         filtersDF = list(),
         filtersSE = list(),
         imputeMethod = "MinProb",
+        imputeArgs = list(),
         assaysForExport = NULL,
         addAbundanceValues = TRUE,
         addHeatmaps = TRUE,
@@ -348,6 +349,12 @@ test_that("runSpectronautAnalysis works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(runSpectronautAnalysis, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## imputeArgs
+    args <- args0
+    args$imputeArgs <- 1
+    expect_error(do.call(runSpectronautAnalysis, args),
+                 "'imputeArgs' must be of class 'list'")
 
     ## assaysForExport
     args <- args0